refactor: reduce STM8 firmware footprint

+++ firmware/gps.h
@@ -54,7 +54,7 @@ uint8_t gps_save_settings(void);
 uint8_t gps_get_fix(struct gps_fix *fix);
 uint8_t gps_wake_up(void);
 void gps_startup_delay(void);
-void delay_ms(unsigned long ms);
+void delay_ms(uint16_t ms);
 void uart_power_save(uint8_t on);
